backend_app/app/requirements.txt
fastapi
gunicorn
sqlalchemy
psycopg2-binary
python-multipart
pydantic-settings
pydantic[email]
# For JWT auth
passlib[bcrypt]
pyjwt
# Testing and development
pytest
bcrypt==4.2.1
# Azure
aiohttp
azure-identity==1.16.1
azure-core==1.32.0
azure-cosmos==4.8.0
azure-common==1.1.28
azure-storage-blob
httpx
uvicorn[standard]==0.32.0
requests==2.32.3
python-dotenv==1.0.1
python-jose[cryptography]==3.3.0
PyJWT[crypto]>=2.8.0
cryptography>=41.0.0